+++ c7n/resources/kms.py
@@ -1,5 +1,6 @@
 # Copyright The Cloud Custodian Authors.
 # SPDX-License-Identifier: Apache-2.0
+from c7n.filters.iamaccess import _account, PolicyChecker
 from botocore.exceptions import ClientError
 
 from datetime import datetime, timezone
@@ -176,6 +177,21 @@ class KeyRotationStatus(ValueFilter):
                 r.get('KeyRotationEnabled', {}))]
 
 
+class KMSPolicyChecker(PolicyChecker):
+    # https://docs.aws.amazon.com/kms/latest/developerguide/policy-conditions.html#conditions-kms
+
+    def handle_kms_calleraccount(self, s, c):
+        return bool(set(map(_account, c['values'])).difference(self.allowed_accounts))
+
+    def handle_kms_viaservice(self, s, c):
+        # We dont filter on service so all are presumed allowed
+        return False
+
+    def handle_kms_grantoperations(self, s, c):
+        # We dont filter on GrantOperations so all are presumed allowed
+        return False
+
+
 @Key.filter_registry.register('cross-account')
 @KeyAlias.filter_registry.register('cross-account')
 class KMSCrossAccountAccessFilter(CrossAccountAccessFilter):
@@ -192,6 +208,8 @@ class KMSCrossAccountAccessFilter(CrossAccountAccessFilter):
                   - type: cross-account
     """
     permissions = ('kms:GetKeyPolicy',)
+
+    checker_factory = KMSPolicyChecker
 
     def process(self, resources, event=None):
         client = local_session(
