+++ tests/data/placebo/test_cross_account_kms_condition_keys_2/kms.CreateKey_1.json
@@ -1,0 +1,33 @@
+{
+    "status_code": 200,
+    "data": {
+        "KeyMetadata": {
+            "AWSAccountId": "644160558196",
+            "KeyId": "ba74e5eb-e6f6-4a7c-9732-8c49f4d9caa7",
+            "Arn": "arn:aws:kms:af-south-1:644160558196:key/ba74e5eb-e6f6-4a7c-9732-8c49f4d9caa7",
+            "CreationDate": {
+                "__class__": "datetime",
+                "year": 2022,
+                "month": 8,
+                "day": 22,
+                "hour": 10,
+                "minute": 25,
+                "second": 26,
+                "microsecond": 681000
+            },
+            "Enabled": true,
+            "Description": "test-cross-account-3",
+            "KeyUsage": "ENCRYPT_DECRYPT",
+            "KeyState": "Enabled",
+            "Origin": "AWS_KMS",
+            "KeyManager": "CUSTOMER",
+            "CustomerMasterKeySpec": "SYMMETRIC_DEFAULT",
+            "KeySpec": "SYMMETRIC_DEFAULT",
+            "EncryptionAlgorithms": [
+                "SYMMETRIC_DEFAULT"
+            ],
+            "MultiRegion": false
+        },
+        "ResponseMetadata": {}
+    }
+}
