+++ tests/data/placebo/test_cross_account_kms_condition_keys_2/kms.DisableKey_1.json
@@ -1,0 +1,6 @@
+{
+    "status_code": 200,
+    "data": {
+        "ResponseMetadata": {}
+    }
+}
